Someone in the carddb had a card db thats longer than 14 chars, and has a name with non-ascii characters in it.
When we try to update the carddb with those details the long card id
make us print an error, which then throws a unicode exception and prevents
us from updating the carddb.

It looks as if sys.stdout.encoding is None when stdout is a pipe or file,
so print trys to convert unicode things into ascii, which does not work.

We can fiddle with sys.stdout to try to make it work, and it works on the
command line, but does not work in the test suite.

So bodge this by having the user model only output unicode when sys.stdout
is not None.

Also wrap some stuff in transaction.atomic() - tho we don't need it since
we're now testing for len(cardid) <= 14