Add optional JSON responses for ACNode endpoints
The numeric responses have caused a lot of parsing errors
on the ACNodes, sometimes resulting in false door triggering.
This commit adds optional JSON outputs, which should be less error
prone.

It uses the new header, X-AC-JSON to trigger this mode. The reasoning
behind this is that using Accept could trigger ambiguity on a node
(or browser) supporting either mode.

A subset of the tests are duplicated as JSON versions are duplicated to check
the functionality. Once JSON becomes the primary way of communicating with nodes,
the main tests should be updated instead.