Fix permissions dropdowns for maintainers
A previous commit added support for searchable dropdowns for adding
users to tools. This seems like a great idea in theory, but in practice,
the implementation was slightly flawed.

The problem was that non-admin maintainers did not have view access to
the user list, so the autocomplete dropdown did not load. This commit
fixes that by giving tool maintainers view access to the user list.